+++ doc/CHANGES.md
@@ -62,6 +62,7 @@ Change log
 ## 4.2.5-dev (TBD)
 
 * fix [#1784](https://github.com/gridstack/gridstack.js/issues/1784) `removable:true` working by itself (without needing `acceptWidgets:true`)
+* fix [#1791](https://github.com/gridstack/gridstack.js/pull/1791) removed drag flicker and scroll issue. Thanks [@nelsieborja](https://github.com/nelsieborja)
 
 ## 4.2.5 (2021-5-31)