+++ doc/CHANGES.md
@@ -63,6 +63,7 @@ Change log
 
 * fix [#1784](https://github.com/gridstack/gridstack.js/issues/1784) `removable:true` working by itself (without needing `acceptWidgets:true`)
 * fix [#1791](https://github.com/gridstack/gridstack.js/pull/1791) removed drag flicker and scroll issue. Thanks [@nelsieborja](https://github.com/nelsieborja)
+* better doc for save [#1795](https://github.com/gridstack/gridstack.js/issues/1795)
 
 ## 4.2.5 (2021-5-31)